[SubshapeReference] Harden references against non-deterministic OCCT subshape order (#105)
Replace raw-index fallbacks for copy and boolean-derived subshapes with
purely topological composite references, so references stay valid when the
face/edge enumeration order changes (as it does in OCCT 8).

+++ Tests/Test.Utils/Utils/TestData.cs
@@ -199,10 +199,10 @@ public static class TestData
 
     //--------------------------------------------------------------------------------------------------
 
-    public static Guid CreateGuid(byte no)
+    public static Guid CreateGuid(int no)
     {
         Span<byte> bytes = stackalloc byte[16];
-        bytes[15] = no;
+        bytes[15] = (byte)no;
         return new Guid(bytes);
     }
